requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
psycopg2-binary>=2.9.6
openpyxl>=3.1.2
Pillow>=10.0.0